+++ plutus-extra/src/Plutus/Contract/Test/Extra.hs
@@ -10,6 +10,7 @@ module Plutus.Contract.Test.Extra (
   valueAtComputedAddress,
   dataAtComputedAddress,
   dataAtComputedAddressWithState,
+  valueAtComputedAddressWithState,
   utxoAtComputedAddressWithState,
 ) where
 
@@ -32,7 +33,7 @@ import Data.List (foldl')
 import Data.Map qualified as Map
 import Data.Maybe (mapMaybe)
 import Data.Row (Row)
-import Data.Text.Prettyprint.Doc (Doc, Pretty, colon, indent, line, pretty, viaShow, vsep, (<+>))
+import Data.Text.Prettyprint.Doc (Doc, Pretty, align, colon, indent, pretty, viaShow, vsep, (<+>))
 import Data.Void (Void)
 import Prelude
 
@@ -170,6 +171,36 @@ walletFundsChangeWithAccumStateImpl exact contract inst wallet toDlt =
                  ]
       pure result
 
+data CheckedState = CheckedState
+  { checkedAddress :: Ledger.Address
+  , checkedData :: [CheckedData]
+  }
+
+data CheckedData where
+  CheckedDatas :: UtxoMap -> CheckedData
+  CheckedValue :: Ledger.Value -> CheckedData
+  CheckedUtxos :: UtxoMap -> CheckedData
+  CheckedWriter :: forall w. Pretty w => w -> CheckedData
+
+instance Pretty CheckedState where
+  pretty (CheckedState address datas) =
+    "At the address" <+> pretty address <+> ":"
+      <+> align (vsep $ map pretty datas)
+
+instance Pretty CheckedData where
+  pretty (CheckedDatas utxos) =
+    "Data was"
+      <+> foldMap (foldMap pretty . Ledger.txData . Ledger.txOutTxTx) utxos
+  pretty (CheckedValue value) =
+    "Funds was"
+      <+> pretty value
+  pretty (CheckedUtxos utxos) =
+    "UTxO was"
+      <+> foldMap viaShow utxos
+  pretty (CheckedWriter w) =
+    "Contract writer data was"
+      <+> pretty w
+
 {- | Check that the funds at a computed address meet some condition.
  The address is computed using data acquired from contract's writer instance.
 
@@ -196,15 +227,52 @@ valueAtComputedAddress ::
   (Ledger.Value -> Bool) ->
   TracePredicate
 valueAtComputedAddress contract inst addressGetter check =
-  utxoAtComputedAddress contract inst addressGetter $ \addr utxoMap -> do
+  utxoAtComputedAddress contract inst addressGetter $ \_ addr utxoMap ->
     let value = foldMap (Ledger.txOutValue . Ledger.txOutTxOut) utxoMap
-        result = check value
-    unless result $
-      tell @(Doc Void) ("Funds at address" <+> pretty addr <+> "were" <> pretty value)
-    return result
+     in showStateIfFailAndReturn
+          [CheckedValue value]
+          addr
+          (check value)
+
+{- | Check that the funds at a computed address
+ and data aquired from contract's writer instance meet some condition.
+ The address is computed using data acquired from contract's writer instance.
+
+  @since 2.2
+-}
+valueAtComputedAddressWithState ::
+  forall
+    (w :: Type)
+    (s :: Row Type)
+    (e :: Type)
+    (a :: Type)
+    (contract :: Type -> Row Type -> Type -> Type -> Type).
+  ( Monoid w
+  , Pretty w
+  , IsContract contract
+  ) =>
+  -- | The 'IsContract' code
+  contract w s e a ->
+  -- | The 'ContractInstanceTag', acquired inside the
+  -- 'Plutus.Trace.Emulator.EmulatorTrace'
+  ContractInstanceTag ->
+  -- | The function computing 'Ledger.Address'
+  (w -> Maybe Ledger.Address) ->
+  -- | The @datum@ predicate
+  (w -> Ledger.Value -> Bool) ->
+  TracePredicate
+valueAtComputedAddressWithState contract inst addressGetter check =
+  utxoAtComputedAddress contract inst addressGetter $ \w addr utxoMap ->
+    let value = foldMap (Ledger.txOutValue . Ledger.txOutTxOut) utxoMap
+     in showStateIfFailAndReturn
+          [CheckedValue value, CheckedWriter w]
+          addr
+          (check w value)
 
 {- | Check that the datum at a computed address meet some condition.
  The address is computed using data acquired from contract's writer instance.
+
+  @since 1.1
 -}
 dataAtComputedAddress ::
   forall
@@ -229,19 +297,18 @@ dataAtComputedAddress ::
   (datum -> Bool) ->
   TracePredicate
 dataAtComputedAddress contract inst addressGetter check =
-  utxoAtComputedAddress contract inst addressGetter $ \addr utxoMap -> do
+  utxoAtComputedAddress contract inst addressGetter $ \_ addr utxoMap ->
     let datums = mapMaybe (uncurry $ getTxOutDatum @datum) $ Map.toList utxoMap
-        result = any check datums
-    unless result $
-      tell @(Doc Void)
-        ( "Data at address" <+> pretty addr <+> "was"
-            <+> foldMap (foldMap pretty . Ledger.txData . Ledger.txOutTxTx) utxoMap
-        )
-    return result
+     in showStateIfFailAndReturn
+          [CheckedDatas utxoMap]
+          addr
+          (any check datums)
 
 {- | Check that the datum at a computed address
  and data aquired from contract's writer instance meet some condition.
  The address is computed using data acquired from contract's writer instance.
+
+  @since 2.1
 -}
 dataAtComputedAddressWithState ::
   forall
@@ -267,24 +334,58 @@ dataAtComputedAddressWithState ::
   (w -> datum -> Bool) ->
   TracePredicate
 dataAtComputedAddressWithState contract inst addressGetter check =
-  utxoAtComputedAddressWithStateImpl contract inst addressGetter $ \w addr utxoMap -> do
+  utxoAtComputedAddress contract inst addressGetter $ \w addr utxoMap ->
     let datums = mapMaybe (uncurry $ getTxOutDatum @datum) $ Map.toList utxoMap
-        result = any (check w) datums :: Bool
-    unless result $
-      tell @(Doc Void)
-        ( "Data at address" <+> pretty addr <+> "was"
-            <+> foldMap (foldMap pretty . Ledger.txData . Ledger.txOutTxTx) utxoMap
-            <> line
-            <> "Contract writer data was"
-            <+> pretty w
-        )
-    return result
+     in showStateIfFailAndReturn
+          [CheckedDatas utxoMap, CheckedWriter w]
+          addr
+          (any (check w) datums)
 
-{- | Extract UTxOs at a computed address and call continuation returning
- Boolean value based on both the address and UTxOs.
+showStateIfFailAndReturn ::
+  forall effs.
+  Member (Writer (Doc Void)) effs =>
+  [CheckedData] ->
+  Ledger.Address ->
+  Bool ->
+  Eff effs Bool
+showStateIfFailAndReturn datas addr result = do
+  unless result $ tell @(Doc Void) $ pretty $ CheckedState addr datas
+  pure result
+
+{- | Check that the UTxO at a computed address
+ and data aquired from contract's writer instance meet some condition.
  The address is computed using data acquired from contract's writer instance.
 
-  @since 1.1
+  @since 2.1
+-}
+utxoAtComputedAddressWithState ::
+  forall
+    (w :: Type)
+    (s :: Row Type)
+    (e :: Type)
+    (a :: Type)
+    (contract :: Type -> Row Type -> Type -> Type -> Type).
+  ( Monoid w
+  , Pretty w
+  , IsContract contract
+  ) =>
+  contract w s e a ->
+  ContractInstanceTag ->
+  (w -> Maybe Ledger.Address) ->
+  (w -> UtxoMap -> Bool) ->
+  TracePredicate
+utxoAtComputedAddressWithState contract inst getter check =
+  utxoAtComputedAddress contract inst getter $ \w addr utxoMap ->
+    showStateIfFailAndReturn
+      [CheckedUtxos utxoMap, CheckedWriter w]
+      addr
+      (check w utxoMap)
+
+{- | Extract UTxOs at a computed address and call continuation returning
+ Boolean value based the address, UTxOs, and data aquired from contract's writer instance.
+ The address is computed using data acquired from contract's writer instance.
+
+   @since 2.1
 -}
 utxoAtComputedAddress ::
   forall
@@ -307,77 +408,9 @@ utxoAtComputedAddress ::
   -- | The function computing 'Ledger.Address'
   (w -> Maybe Ledger.Address) ->
   -- | The continuation function acting as a predicate
-  (Ledger.Address -> UtxoMap -> Eff effs Bool) ->
-  Folds.EmulatorEventFoldM effs Bool
-utxoAtComputedAddress contract inst addressGetter cont =
-  utxoAtComputedAddressWithStateImpl contract inst addressGetter (const cont)
-
-{- | Check that the UTxO at a computed address
- and data aquired from contract's writer instance meet some condition.
- The address is computed using data acquired from contract's writer instance.
-
-  @since 1.1
--}
-utxoAtComputedAddressWithState ::
-  forall
-    (w :: Type)
-    (s :: Row Type)
-    (e :: Type)
-    (a :: Type)
-    (contract :: Type -> Row Type -> Type -> Type -> Type).
-  ( Monoid w
-  , Pretty w
-  , IsContract contract
-  ) =>
-  contract w s e a ->
-  ContractInstanceTag ->
-  (w -> Maybe Ledger.Address) ->
-  (w -> UtxoMap -> Bool) ->
-  TracePredicate
-utxoAtComputedAddressWithState contract inst getter check =
-  utxoAtComputedAddressWithStateImpl contract inst getter $ \w addr utxoMap ->
-    let result = check w utxoMap
-     in do
-          unless result $
-            tell @(Doc Void)
-              ( "UTxO at address" <+> pretty addr <+> "was"
-                  <+> foldMap viaShow utxoMap
-                  <> line
-                  <> "Contract writer data was"
-                  <+> pretty w
-              )
-          return result
-
-{-utxoAtComputedAddressWithStateImpl contract tag getter
-$ \w addr -> return . predicate w addr-}
-
-{- | Similar to 'utxoAtComputedAddress' but continuation have access
- to a data aquired from contract's writer instance.
--}
-utxoAtComputedAddressWithStateImpl ::
-  forall
-    (effs :: [Type -> Type])
-    (w :: Type)
-    (s :: Row Type)
-    (e :: Type)
-    (a :: Type)
-    (contract :: Type -> Row Type -> Type -> Type -> Type).
-  ( Member (Error Folds.EmulatorFoldErr) effs
-  , Member (Writer (Doc Void)) effs
-  , Monoid w
-  , IsContract contract
-  ) =>
-  -- | The 'IsContract' code
-  contract w s e a ->
-  -- | The 'ContractInstanceTag', acquired inside the
-  -- 'Plutus.Trace.Emulator.EmulatorTrace'
-  ContractInstanceTag ->
-  -- | The function computing 'Ledger.Address'
-  (w -> Maybe Ledger.Address) ->
-  -- | The continuation function acting as a predicate
   (w -> Ledger.Address -> UtxoMap -> Eff effs Bool) ->
   Folds.EmulatorEventFoldM effs Bool
-utxoAtComputedAddressWithStateImpl contract inst addressGetter cont =
+utxoAtComputedAddress contract inst addressGetter cont =
   flip
     postMapM
     ( (,)
@@ -388,7 +421,7 @@ utxoAtComputedAddressWithStateImpl contract inst addressGetter cont =
       case addressGetter w of
         Nothing -> do
           tell @(Doc Void) $ "Could not compute address using the given getter"
-          return False
+          pure False
         Just addr -> do
           let step = \case
                 TxnValidate _ txn _ -> AM.updateAddresses (Ledger.Valid txn)