+++ tasty-plutus/src/Test/Tasty/Plutus/Context.hs
@@ -45,8 +45,8 @@ module Test.Tasty.Plutus.Context (
   paysToWallet,
   paysLovelaceToPubKey,
   paysLovelaceToWallet,
-  paysSelf,
-  paysOther,
+  paysToSelf,
+  paysToOther,
 
   -- ** Spending
   spendsFromPubKey,
@@ -172,30 +172,30 @@ paysToWallet wallet = paysToPubKey (walletPubKeyHash wallet)
 
 {- | Indicate that the script being tested must pay itself the given amount.
 
- @since 1.0
+ @since 4.0
 -}
-paysSelf ::
+paysToSelf ::
   forall (p :: Purpose) (a :: Type).
   (ToData a) =>
   Value ->
   a ->
   ContextBuilder p
-paysSelf v dt =
+paysToSelf v dt =
   output . Output (OwnType . toBuiltinData $ dt) $ v
 
 {- | Indicate that the script being tested must pay another script the given
  amount.
 
- @since 1.0
+ @since 4.0
 -}
-paysOther ::
+paysToOther ::
   forall (p :: Purpose) (a :: Type).
   (ToData a) =>
   ValidatorHash ->
   Value ->
   a ->
   ContextBuilder p
-paysOther hash v dt =
+paysToOther hash v dt =
   output . Output (ScriptType hash . toBuiltinData $ dt) $ v
 
 {- | As 'paysToPubKey', but using Lovelace.